PDF_File_Analysis_BE/feature_extractor/requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
colorama==0.3.7
future==0.18.3
jsbeautifier==1.6.2
marshmallow==3.20.1
packaging==23.2
peepdf==0.4.2
Pillow==3.2.0
PyPDF2==3.0.1
pythonaes==1.0
six==1.16.0
soupsieve==2.5
timeout-decorator==0.5.0
yara-python==4.3.1
